pytest fixes: moved funcs and added __init__.py to test dirs with common.py
