+++ holopy/core/tests/meson.build
@@ -1,0 +1,6 @@
+core_test_sources = [
+'__init__.py',
+'common.py'
+]
+
+py.install_sources(core_test_sources, subdir: 'holopy/core/tests')
