+++ holopy/scattering/meson.build
@@ -10,3 +10,4 @@ py.install_sources(scattering_sources, subdir: 'holopy/scattering')
 subdir('scatterer')
 subdir('theory')
 subdir('third_party')
+subdir('tests')
