+++ holopy/scattering/tests/meson.build
@@ -1,0 +1,6 @@
+scattering_test_sources = [
+'__init__.py',
+'common.py'
+]
+
+py.install_sources(scattering_test_sources, subdir: 'holopy/scattering/tests')
