+++ holopy/scattering/tests/test_scatteringtheory.py
@@ -13,7 +13,7 @@ from holopy.scattering.scatterer import Sphere, Spheres, Ellipsoid
 from holopy.scattering.errors import TheoryNotCompatibleError
 from holopy.scattering.interface import prep_schema
 from holopy.scattering.tests.common import xschema as XSCHEMA
-
+from holopy.scattering.tests.common import MockTheory, MockScatteringMatrixBasedTheory
 
 SPHERE = Sphere(n=1.5, r=1.0, center=(0, 0, 2))
 SPHERES = Spheres([
@@ -28,36 +28,6 @@ MEDTOLS = {'atol': 1e-7, 'rtol': 1e-7}
 SCAT_SCHEMA = prep_schema(
     detector_grid(shape=(5, 5), spacing=.1),
     medium_index=1.33, illum_wavelen=0.66, illum_polarization=False)
-
-
-class MockTheory(ScatteringTheory):
-    """Minimally-functional daughter of ScatteringTheory for fast tests."""
-    def __init__(*args, **kwargs):
-        pass  # an init is necessary for the repr
-
-    def can_handle(self, scatterer):
-        return isinstance(scatterer, Sphere)
-
-    def raw_fields(self, positions, *args, **kwargs):
-        return np.ones(positions.shape, dtype='complex128')
-
-
-class MockScatteringMatrixBasedTheory(ScatteringTheory):
-    """Minimally-functional daughter of ScatteringTheory which
-    uses the scattering matrix pathway, for fast tests.
-    Smells like a Rayleigh scatterer, just for fun. But it's not"""
-    def __init__(*args, **kwargs):
-        pass  # an init is necessary for the repr
-
-    def can_handle(self, scatterer):
-        return isinstance(scatterer, Sphere)
-
-    def raw_scat_matrs(self, scatterer, positions, *args, **kwargs):
-        strength = scatterer.n * scatterer.r
-        scattering_matrix = np.array(
-            [np.eye(2) for _ in range(positions.shape[1])])
-        return strength * scattering_matrix.astype('complex128')
-
 
 class TestScatteringTheory(unittest.TestCase):
     @attr('fast')
